refactor: merge thumbnail data sources
